+++ src/Authentication/Adapter/TokenHeaderAuthenticationAdapter.php
@@ -1,0 +1,153 @@
+<?php
+/**
+ * Abstarct header authentication adapter file
+ *
+ * @copyright Copyright (c) 2016, final gene <info@final-gene.de>
+ * @author    Frank Giesecke <frank.giesecke@final-gene.de>
+ */
+
+namespace FinalGene\RestResourceAuthenticationModule\Authentication\Adapter;
+
+use FinalGene\RestResourceAuthenticationModule\Exception\TokenException;
+use FinalGene\RestResourceAuthenticationModule\Exception\IdentityNotFoundException;
+use FinalGene\RestResourceAuthenticationModule\Service\IdentityServiceInterface;
+use Zend\Authentication\Result;
+use Zend\Http\Request;
+
+/**
+ * Class TokenHeaderAuthenticationAdapter
+ *
+ * @package FinalGene\RestResourceAuthenticationModule\Authentication\Adapter
+ */
+class TokenHeaderAuthenticationAdapter extends AbstractHeaderAuthenticationAdapter
+{
+    const AUTH_HEADER = 'Authorization';
+    const AUTH_IDENTIFIER = 'Token';
+
+    /**
+     * @var IdentityServiceInterface
+     */
+    private $identityService;
+
+    /**
+     * Get $identityService
+     *
+     * @return IdentityServiceInterface
+     */
+    public function getIdentityService()
+    {
+        return $this->identityService;
+    }
+
+    /**
+     * @param IdentityServiceInterface $identityService
+     * @return TokenHeaderAuthenticationAdapter
+     */
+    public function setIdentityService(IdentityServiceInterface $identityService)
+    {
+        $this->identityService = $identityService;
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function authenticate()
+    {
+        $request = $this->getRequest();
+        $header = $request->getHeaders();
+
+        if (!$header->has(self::AUTH_HEADER)) {
+            return $this->buildErrorResult('Authorization header missing');
+        }
+
+        $authorization = $header->get(self::AUTH_HEADER)->getFieldValue();
+        if (0 !== strpos($authorization, self::AUTH_IDENTIFIER . ' ')) {
+            return $this->buildErrorResult('Invalid authorization header');
+        }
+
+        try {
+            $publicKey = $this->extractPublicKey($authorization);
+            $signature = $this->extractSignature($authorization);
+            $secret = $this->getIdentityService()->getSecret($publicKey);
+
+        } catch (TokenException $e) {
+            return $this->buildErrorResult($e->getMessage(), $e->getCode());
+
+        } catch (IdentityNotFoundException $e) {
+            return $this->buildErrorResult($e->getMessage(), Result::FAILURE_IDENTITY_NOT_FOUND);
+        }
+
+        $hmac = $this->getHmac($request, $secret);
+        if ($hmac !== $signature) {
+            if ($header->has('XDEBUG_SESSION_START')) {
+                error_log(sprintf('Signature for identity `%s`: %s', $publicKey, $hmac));
+            }
+            return $this->buildErrorResult('Signature does not match', Result::FAILURE_CREDENTIAL_INVALID);
+        }
+
+        return new Result(Result::SUCCESS, null);
+    }
+
+    /**
+     * Extract public key from authorization
+     *
+     * @param $authorization
+     *
+     * @return string
+     * @throws TokenException
+     */
+    protected function extractPublicKey($authorization)
+    {
+        $identifierLength = strlen(self::AUTH_IDENTIFIER) + 1;
+        $publicKey = substr(
+            $authorization,
+            $identifierLength,
+            strpos($authorization, ':') - $identifierLength
+        );
+        if (empty($publicKey)) {
+            throw new TokenException(
+                'Public kex not found',
+                Result::FAILURE_IDENTITY_NOT_FOUND
+            );
+        }
+
+        return $publicKey;
+    }
+
+    /**
+     * Extract signature from authorization
+     *
+     * @param $authorization
+     *
+     * @return string
+     * @throws TokenException
+     */
+    protected function extractSignature($authorization)
+    {
+        $signatureStart = strpos($authorization, ':');
+        if (false === $signatureStart) {
+            throw new TokenException(
+                'Signature not found',
+                Result::FAILURE_CREDENTIAL_INVALID
+            );
+        }
+
+        return substr($authorization, $signatureStart + 1);
+    }
+
+    /**
+     * Calculate HMAC for request
+     *
+     * @param Request $request
+     * @param string $secret
+     *
+     * @return string
+     */
+    protected function getHmac(Request $request, $secret)
+    {
+        // Remove authorizatin header to build valid signature
+        $request->getHeaders()->removeHeader($request->getHeader(self::AUTH_HEADER));
+        return hash_hmac('sha256', $request->toString(), $secret);
+    }
+}