+++ src/Exception/AuthenticationException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Authentication service file
+ * Authentication exception file
  *
  * @copyright Copyright (c) 2016, final gene <info@final-gene.de>
  * @author    Frank Giesecke <frank.giesecke@final-gene.de>