+++ src/Exception/TokenException.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Token exception file
+ *
+ * @copyright Copyright (c) 2016, final gene <info@final-gene.de>
+ * @author    Frank Giesecke <frank.giesecke@final-gene.de>
+ */
+
+namespace FinalGene\RestResourceAuthenticationModule\Exception;
+
+/**
+ * Class TokenException
+ *
+ * @package FinalGene\RestResourceAuthenticationModule\Exception
+ */
+class TokenException extends \Exception
+{
+
+}