+++ tests/src/Unit/Authentication/Adapter/AbstractHeaderAuthenticationAdapterTest.php
@@ -37,14 +37,51 @@ class AbstractHeaderAuthenticationAdapterTest extends \PHPUnit_Framework_TestCas
 
     /**
      * @covers FinalGene\RestResourceAuthenticationModule\Authentication\Adapter\AbstractHeaderAuthenticationAdapter::authenticate
+     * @uses FinalGene\RestResourceAuthenticationModule\Authentication\Adapter\AbstractHeaderAuthenticationAdapter::buildErrorResult
      */
     public function testAuthenticate()
     {
-        $adapter = $this->getMockForAbstractClass(AbstractHeaderAuthenticationAdapter::class);
+        $resultMock = $this->getMockBuilder(Result::class)
+            ->setMethods([
+                'getCode'
+            ])
+            ->disableOriginalConstructor()
+            ->getMockForAbstractClass();
+        $resultMock
+            ->expects($this->once())
+            ->method('getCode')
+            ->willReturn(Result::FAILURE_UNCATEGORIZED);
+
+        $adapter = $this->getMockBuilder(AbstractHeaderAuthenticationAdapter::class)
+            ->setMethods([
+                'buildErrorResult'
+            ])
+            ->getMockForAbstractClass();
+        $adapter
+            ->expects($this->once())
+            ->method('buildErrorResult')
+            ->with('No authentication implemented', Result::FAILURE_UNCATEGORIZED)
+            ->willReturn($resultMock);
         /** @var AbstractHeaderAuthenticationAdapter $adapter */
 
         $result = $adapter->authenticate();
         $this->assertInstanceOf(Result::class, $result);
         $this->assertEquals(Result::FAILURE_UNCATEGORIZED, $result->getCode());
+    }
+
+    /**
+     * @covers FinalGene\RestResourceAuthenticationModule\Authentication\Adapter\AbstractHeaderAuthenticationAdapter::buildErrorResult
+     */
+    public function testBuildErrorResult()
+    {
+        $reflection = new \ReflectionClass(AbstractHeaderAuthenticationAdapter::class);
+        $buildErrorResult = $reflection->getMethod('buildErrorResult');
+        $buildErrorResult->setAccessible(true);
+
+        $adapter = $this->getMockForAbstractClass(AbstractHeaderAuthenticationAdapter::class);
+        $result = $buildErrorResult->invokeArgs($adapter, ['']);
+        /** @var Result $result */
+
+        $this->assertInstanceOf(Result::class, $result);
     }
 }