+++ CHANGELOG.md
@@ -2,11 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0] - YYYY-MM-DD
+## [1.2.0] - 2016-01-19
 ### Added
-- Resource listener
-- Service
-- Service initializer
+- Token header authentication adapter
 
 ### Changed 
 - Nothing
@@ -22,3 +20,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Security
 - Nothing
+
+## [1.1.0] - 2016-01-16
+### Added
+- Abstract header authentication adapter
+
+## [1.0.0] - 2016-01-15
+### Added
+- Resource listener
+- Service
+- Service initializer